pyproject.toml
[tool.poetry]
name = "booksum"
version = "0.1.0"
description = ""
authors = ["Ricardo G. Sousa <ricardo@rsousa.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
dvc = "~3.53.0"
dvc-gdrive = "~3.0.1"
nltk = "~3.8.1"
gutenbergpy = "~0.3.5"
pandas = "~2.2.2"
pyarrow = "~17.0.0"
bertopic = "~0.16.3"
numpy = ">=1.26.4,<1.27.0"
xformers = "~0.0.27.post2"
llama-index-llms-groq = "~0.1.4"
llama-index-embeddings-huggingface = "~0.2.2"
evaluate = "~0.4.2"
rouge-score = "~0.1.2"
streamlit = "~1.37.1"
fastapi = "~0.112.0"
uvicorn = "~0.30.5"


[tool.poetry.group.dev.dependencies]
notebook = "~7.2.1"
ipywidgets = "~8.1.3"
wordcloud = "~1.9.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
